Fixed validation rules for Product and Service

+++ app/Http/Requests/Twill/ProductRequest.php
@@ -11,8 +11,8 @@ class ProductRequest extends Request
         return [
             'type' => ['required', 'string', 'max:255'],
             'color' => ['required', 'string', 'max:100'],
-            'size' => ['required', 'integer', 'min:1', 'max:10000'],
-            'price' => ['required', 'integer', 'min:1', 'max:10000000'],
+            'size' => ['required', 'numeric', 'min:1', 'max:10000'],
+            'price' => ['required', 'numeric', 'min:1', 'max:10000000'],
             'quantity' => ['required', 'integer', 'min:1', 'max:100000'],
             'description' => ['nullable', 'string'],
         ];
@@ -23,8 +23,8 @@ class ProductRequest extends Request
         return [
             'type' => ['required', 'string', 'max:255'],
             'color' => ['required', 'string', 'max:100'],
-            'size' => ['required', 'integer', 'min:1', 'max:10000'],
-            'price' => ['required', 'integer', 'min:1', 'max:10000000'],
+            'size' => ['required', 'numeric', 'min:1', 'max:10000'],
+            'price' => ['required', 'numeric', 'min:1', 'max:10000000'],
             'quantity' => ['required', 'integer', 'min:1', 'max:100000'],
             'description' => ['nullable', 'string'],
         ];
